Added Review entity, created bases for DTOs for currently needed entities, created basis for mappers, created service interfaces with methods according to the current specification.

+++ impulse-backend/src/main/java/com/baseddevs/impulsebackend/repository/ShippingRepository.java
@@ -4,5 +4,5 @@ import com.baseddevs.impulsebackend.model.Shipping;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ShippingRepository extends JpaRepository<Shipping, Integer>, JpaSpecificationExecutor<Shipping> {
+public interface ShippingRepository extends JpaRepository<Shipping, Long>, JpaSpecificationExecutor<Shipping> {
 }
